breakout function to load imagen from a checkpoint file, so checkpoints can be passed around and the public can keep finetuning on their own data
