+++ imagen_pytorch/__init__.py
@@ -10,3 +10,7 @@ from imagen_pytorch.elucidated_imagen import ElucidatedImagen
 # config driven creation of imagen instances
 
 from imagen_pytorch.configs import UnetConfig, ImagenConfig, ElucidatedImagenConfig, ImagenTrainerConfig
+
+# utils
+
+from imagen_pytorch.utils import load_imagen_from_checkpoint
