+++ imagen_pytorch/version.py
@@ -1,1 +1,1 @@
-__version__ = '1.0.0'
+__version__ = '1.0.1'
